chore/PLAYER-4980 one more portion of components got compatible with air-bnb eslint
